chore: move handleSelectInputChange to be isnide component
